requirements.txt
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
openpyxl==3.0.5
PyMySQL==0.10.1
python-dotenv==0.15.0
Werkzeug==1.0.1
